move `selector-parser` from `./src/compat` to just `./src`
