+++ packages/tailwindcss/src/signatures.ts
@@ -1,8 +1,8 @@
 import { substituteAtApply } from './apply'
 import { atRule, styleRule, toCss, walk, type AstNode } from './ast'
 import { printArbitraryValue } from './candidate'
-import * as SelectorParser from './compat/selector-parser'
 import { CompileAstFlags, type DesignSystem } from './design-system'
+import * as SelectorParser from './selector-parser'
 import { ThemeOptions } from './theme'
 import { DefaultMap } from './utils/default-map'
 import { dimensions } from './utils/dimensions'
